Features: 	- Merger/Integration of the CashflowPeriod and LossQuadratureMetrics 	- Loss Period Generation and Update Localization into CashFlowPeriod 	- Cash flow Product Fields Comb-down - Pay/Coupon Currency, and MTM'able Period FX Reset Date 	- Period Contract Setup - Incorporation of Period FRI and the Reference Entity Name (both Optional) 	- Categorization into Period Date, Date Generator, "Cash Extensive", and Computed Metric Parameters Groups 	- Period Individual Cash Date Access, Serialization Order, and Set Rationalization 	- Incorporate Stubs for the Period Collateral, Funding, Forward, Credit, and FX Latent States 	- Period-Level Credit Label Incorporation and Accomodation 	- Period-Level Forward Label Incorporation and Accomodation
Bug Fixes:

Samples:
	- Samples Enhancement for Period-Level Floater Credit/Forward Label Incorporation

+++ CreditAnalytics/dev/src/org/drip/analytics/output/ComponentMeasures.java
@@ -180,7 +180,7 @@ public class ComponentMeasures extends org.drip.service.stream.Serializer {
 		if (null == astrField[1] || astrField[1].isEmpty())
 			throw new java.lang.Exception ("ComponentMeasures de-serializer: Cannot locate calc time");
 
-		_dblCalcTime = new java.lang.Double (astrField[1]).doubleValue();
+		_dblCalcTime = new java.lang.Double (astrField[1]);
 
 		if (null == astrField[2] || astrField[2].isEmpty() || NULL_SER_STRING.equalsIgnoreCase
 			(astrField[2]))
